Avoid mouse coordinates misaligned caused by line-prefix
Fixed an issue where the coordinates would shift if the editor was at
the top of the window when using org-indent. The cause is that the
posn-object-x-y function returns the coordinates based on the upper
left of the character string added by line-prefix instead of the upper
left of the image. This is an Emacs issue, but I avoided it by
temporarily disabling line-prefix.