Fix edraw-transform
Shapes like rect and ellipse can now be rotated. However, since it
uses the transform attribute, it also has its drawbacks. Once a shape
has the transform attribute, simple translations will also be done
with the transform attribute.

edraw-transform tries to transform by moving the anchor point
coordinates as much as possible. In other words, it tries to preserve
line widths, etc. But if that is not possible edraw-transform uses the
transform property.