Add initial check of SVG tree
edraw-editor checks the given svg tree on initialization.

If the root is not an svg element, issue a warning and use the default SVG.

If the svg element does not have a width or height attribute, issue a warning and apply the default document size. If this countermeasure is not taken, arithmetic operations with nil will cause an error.