Fix error when deleting anchor point other than path
Right-clicking an anchor point does not bring up the context menu, but
you can delete the selected anchor point by pressing the Delete
key. At that time, an error occurs because edraw-delete-point is not
implemented in anything other than edraw-shape-point-path.

If the shape does not support anchor point deletion, an explicit error
message will be displayed. Also, in the case of text, the text itself
will be deleted.