Snap points when moving with arrow keys
When moving a point using the arrow keys, you can now move to the next
grid line.
Added the symbol 'snap' as the movement distance specification value
for the customization variable edraw-editor-move-distance-by-arrow-key.
Snap only when moving a single anchor or handle. It does not snap when
moving the entire shape.