Preserve spaces within text elements
The `space' property has been added to text elements. This is output
as the xml:space attribute of the text element when it is
printed. Since librsvg does not support the `white-space' property, I
decided to use `xml:space'. Also, since libxml-parse-xml-region
discards the XML namespace, I decided to name the property simply
`space' and add xml: when printing with edraw-svg-print.

When adding a new text element, the default value for `xml:space' is
"preserve".