Add color setting functionality by color name to the color picker
This is not a function that makes the color picker output the color
name, but a function that changes the current color of the color
picker by specifying a color name. The color names that can be
specified are not limited to the color names defined by Emacs or
CSS. You can use your favorite color swatches by changing the
edraw-color-read-name-sources variable.