Use edraw-y-or-n-p
Generally, edraw-y-or-n-p is used rather than y-or-n-p. However,
y-or-n-p is used in places where mouse use is almost never used.

Added edraw-y-or-n-p-function variable. Use edraw-y-or-n-p--popup-menu
on Windows. Use y-or-n-p on other occasions. At least on Android,
edraw-y-or-n-p--popup-menu does not work properly.

edraw-y-or-n-p--popup-menu uses edraw-use-dialog-box-p to determine
whether keyboard or mouse operation is used.