Preserve marker size and stroke width
Ensure that changing marker size (markerWidth and markerHeight) does
not affect the stroke width of lines within the marker.  The stroke
width within markers now always matches the stroke-width of the element
that has the marker.

Prevent marker size from changing when the stroke-width of the element
with the marker changes.  (The old behavior can be restored via the
edraw-svg-marker-size-independent-of-stroke-width variable)